requirements.txt
asgiref==3.2.7
autopep8==1.4.4
certifi==2019.11.28
Django==3.0.5
django-crispy-forms==1.9.0
django-taggit==1.2.0
Markdown==3.2.1
Pillow==7.1.1
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.1
wincertstore==0.2
